move back to prompt
